+++ Userland/Libraries/LibJS/Runtime/Intl/RelativeTimeFormatConstructor.h
@@ -27,4 +27,6 @@ private:
     JS_DECLARE_NATIVE_FUNCTION(supported_locales_of);
 };
 
+ThrowCompletionOr<RelativeTimeFormat*> initialize_relative_time_format(GlobalObject& global_object, RelativeTimeFormat& relative_time_format, Value locales_value, Value options_value);
+
 }
